salesflow-ai/1-python-data-engine/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
async-timeout==5.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.121.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.2
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
redis==7.0.1
reportlab==4.4.5
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.38.0
